reformulating order, brand and customer imports due to the restructuring of the service architecture
