+++ src/app/dashboard/dashboard.component.ts
@@ -1,5 +1,5 @@
 import { Component, OnInit } from '@angular/core';
-import { BrandsService, ICategory, IBrand, IOrder, IProduct } from '../services/brands.service';
+import { BrandsService, ICategory, IBrand, IOrder, IProduct } from '../services/brand/brands.service';
 
 interface ChartDataPoint {
   label: string;
@@ -59,9 +59,9 @@ export class DashboardComponent implements OnInit {
   categorySelected(categoryId: number): void {
     this.selectedCategory = categoryId;
     this.brandsService.getProductsByCategoryId(categoryId).subscribe(
-      data => {
+      response => {
         this.selected.category = this.categories.find(c => c.categoryId === categoryId)?.categoryName || '';
-        this.products = data;
+        this.products = response;
       },
       error => console.error(error)
     );
